+++ test/partial_reindex_test.rb
@@ -1,0 +1,27 @@
+require_relative "test_helper"
+
+class PartialReindexTest < Minitest::Test
+  def test_basic
+    store [{name: "Hi", color: "Blue"}]
+
+    # normal search
+    assert_search "hi", ["Hi"], fields: [:name], load: false
+    assert_search "blue", ["Hi"], fields: [:color], load: false
+
+    # update
+    Product.first.update_column :name, "Bye"
+    Product.first.update_column :color, "Red"
+    Product.searchkick_index.refresh
+
+    # index not updated
+    assert_search "hi", ["Hi"], fields: [:name], load: false
+    assert_search "blue", ["Hi"], fields: [:color], load: false
+
+    # partial reindex
+    Product.partial_reindex(:search_name)
+
+    # name updated, but not color
+    assert_search "bye", ["Bye"], fields: [:name], load: false
+    assert_search "blue", ["Bye"], fields: [:color], load: false
+  end
+end